almost done!

+++ v03_pipeline/lib/tasks/base_variant_annotations_table_test.py
@@ -7,30 +7,31 @@ from unittest.mock import patch
 import hail as hl
 import luigi.worker
 
-from v03_pipeline.lib.definitions import DataRoot, DatasetType, Env, ReferenceGenome
+from v03_pipeline.lib.definitions import DatasetType, Env, ReferenceGenome
 from v03_pipeline.lib.tasks.base_variant_annotations_table import (
     BaseVariantAnnotationsTableTask,
 )
-from v03_pipeline.lib.tasks.files import GCSorLocalFolderTarget, GCSorLocalTarget
+from v03_pipeline.lib.tasks.files import GCSorLocalFolderTarget
 
 
-@patch('v03_pipeline.lib.definitions.DataRoot')
+@patch('v03_pipeline.lib.paths.DataRoot')
 class BaseVariantAnnotationsTableTest(unittest.TestCase):
 
     def setUp(self):
         self._temp_dir = tempfile.TemporaryDirectory().name
 
     def tearDown(self):
-        shutil.rmtree(self._temp_dir)
+        if os.path.isdir(self._temp_dir):
+            shutil.rmtree(self._temp_dir)
 
     def test_base_variant_annotations_table(self, mock_dataroot) -> None:
-        mock_dataroot.LOCAL_DATASETS.return_value = self._temp_dir
+        mock_dataroot.TEST_DATASETS.value = self._temp_dir
         vat_task = BaseVariantAnnotationsTableTask(
             env=Env.TEST,
             reference_genome=ReferenceGenome.GRCh38,
             dataset_type=DatasetType.SNV,
         )
-        self.assertEqual(vat_task.output().path, 'seqr-datasets/GRCh38/v03/SNV/annotations.ht')
+        self.assertEqual(vat_task.output().path, f'{self._temp_dir}/GRCh38/v03/SNV/annotations.ht')
         self.assertFalse(vat_task.output().exists())
         self.assertFalse(vat_task.complete())
 
@@ -39,7 +40,7 @@ class BaseVariantAnnotationsTableTest(unittest.TestCase):
         worker.run()
         self.assertTrue(GCSorLocalFolderTarget(vat_task.output().path).exists())
         self.assertTrue(vat_task.complete())
-        
+
         mt = hl.read_matrix_table(vat_task.output().path)
         self.assertEqual(mt.count(), (0, 0))
-        self.assertEqual(mt.row_key.fields, ('locus', 'alleles'))
+        self.assertEqual(mt.row_key.collect(), ('locus', 'alleles'))
